Fix some casting in string-util.h
This was prompted by Coverity 345528, but probably won't solve it.

In C, when you shift a value left, it will be promoted to a larger
integer type. Typically this is int (but if all the values of the
original type won't fit into an int, it'll be an unsigned int).

For the rest of this discussion, imagine we have:

  uint8_t a, b, c, d;

So a << 8 will become an int.

So will a << 16, even if an int happens to only be a 16 bit
type on your platform. Thus data is lost.

So the correct fix is to cast the type to something of at least the
required target size before the shift. For example:

  (uint32_t)a << 24 | (uint32_t)b << 16 | (uint32_t)c << 8 | (uint32_t)d

So that'll be fine on an int = int16_t platform. But on an int = int64_t
platform, the left shifts will promote to ints, and that whole expression
will end up as an int.

So to avoid the possibility of 'implicit narrowing cast' warnings, we
need to cast the whole thing:

 (uint32_t)((uint32_t)a << 24 | (uint32_t)b << 16 |
            (uint32_t)c << 8 | (uint32_t)d)

+++ include/mupdf/fitz/string-util.h
@@ -338,27 +338,27 @@ int fz_toupper(int c);
 
 static inline uint16_t fz_unpack_uint16(const uint8_t *p)
 {
-	return (uint16_t)p[0] << 8 | (uint16_t)p[1];
+	return (uint16_t)((uint16_t)p[0] << 8 | (uint16_t)p[1]);
 }
 
 static inline uint16_t fz_unpack_uint16_le(const uint8_t *p)
 {
-	return (uint16_t)p[1] << 8 | (uint16_t)p[0];
+	return (uint16_t)((uint16_t)p[1] << 8 | (uint16_t)p[0]);
 }
 
 static inline uint32_t fz_unpack_uint32(const uint8_t *p)
 {
-	return (uint32_t)p[0] << 24 | (uint32_t)p[1] << 16 | (uint32_t)p[2] << 8 | (uint32_t)p[3];
+	return (uint32_t)((uint32_t)p[0] << 24 | (uint32_t)p[1] << 16 | (uint32_t)p[2] << 8 | (uint32_t)p[3]);
 }
 
 static inline uint32_t fz_unpack_uint32_le(const uint8_t *p)
 {
-	return (uint32_t)p[3] << 24 | (uint32_t)p[2] << 16 | (uint32_t)p[1] << 8 | (uint32_t)p[0];
+	return (uint32_t)((uint32_t)p[3] << 24 | (uint32_t)p[2] << 16 | (uint32_t)p[1] << 8 | (uint32_t)p[0]);
 }
 
 static inline uint64_t fz_unpack_uint64(const uint8_t *p)
 {
-	return  (
+	return  (uint64_t)(
 		(uint64_t)p[0]<<56 | (uint64_t)p[1]<<48 | (uint64_t)p[2]<<40 | (uint64_t)p[3]<<32 |
 		(uint64_t)p[4]<<24 | (uint64_t)p[5]<<16 | (uint64_t)p[6]<<8 | (uint64_t)p[7]
 	);
@@ -366,7 +366,7 @@ static inline uint64_t fz_unpack_uint64(const uint8_t *p)
 
 static inline uint64_t fz_unpack_uint64_le(const uint8_t *p)
 {
-	return (
+	return (uint64_t)(
 		(uint64_t)p[7]<<56 | (uint64_t)p[6]<<48 | (uint64_t)p[5]<<40 | (uint64_t)p[4]<<32 |
 		(uint64_t)p[3]<<24 | (uint64_t)p[2]<<16 | (uint64_t)p[1]<<8 | (uint64_t)p[0]
 	);
